add a somewhat despicable (but precendented) hack to fix PR4678 & rdar://7309675.
Basically, we want to compile C/ObjC functions like:

void foo() {

into nullary LLVM IR functions.  This is important so we don't get varargs ABI
stuff, and also important because we want to get "arguments are dropped" warnings
from instcombine if you have calls like "foo(42)".

OTOH, we don't want to do this for C++ functions like:

void foo(...) {

While this foo can't do a va_begin, we do want to compile it to a varargs LLVM IR
function.  Not doing so will cause 'arguments are dropped' warning for "foo(42)",
which (unlike the C case) makes perfect sense.

This is a hack because it is predicated on the language kind.  This is because GCC's
tree structure apparently represents these things exactly the same way, so there is
no way to distinguish them.  However, this is a *precedented* hack, because GCC 
already does this in mudflap, fold-const.c etc.

If someone has a better way to get this information, I'm all ears.  Until then, I'm 
just happy to have this old issue fixed. Testcase here: test/FrontendC++/varargs.cpp.