Undo part of my previous commit to mm_malloc.h, going back to the use of stdlib.h. There were numerous problems with forward declaring 'malloc' and 'free', but the most important is that these are reserved by POSIX and may be implemented via a function-like macro.
As suggested by Dale Johannesen, I'm instead guarding the only include of this
in our builtin headers with __STDC_HOSTED__, and I've removed the include of
the header from the test suite. I'll discuss with folks whether we want to have
a hosted section of the test suite or not, and add it (and perhaps other tests)
back there if that's the direction.