Modifications to the exactcheck optimization to address bug #11524.  The issue here was exactcheck2()'s failure to allow a pointer to index out of and then back into its intended referent.
The basic strategy of the fix is to obtain the source pointer from
gepchecks that are being re-written to exactchecks in the exactcheck
optimization pass, and pass that pointer through to exactcheck2()
in a new parameter.  Note that in keeping with the parameter list 
for boundscheck, I inserted the source pointer paramter before 
the base pointer in the exactcheck2() parameter list -- with the 
obvious implications for any code that tries to grab parameters 
from exactcheck2() calls.  

Tested with lit tests in a non-libLTO build.