Merging r168354, r168355 and r168379 into the 3.2 release branch.
Set of  Attributes patches.

Merging r168354:

Make the AttrListPtr object a part of the LLVMContext.

When code deletes the context, the AttributeImpls that the AttrListPtr points to
are now invalid. Therefore, instead of keeping a separate managed static for the
AttrListPtrs that's reference counted, move it into the LLVMContext and delete
it when deleting the AttributeImpls.

Merging r168355:

Merging r168379