Force ffp-contract=off on PowerPC/Linux
In order for the test-suite outputs to match the reference outputs we need to
specify -ffp-contract=off. Not only does extra FMA precision cause differences
between the outputs and the reference outputs, but also between our output and
that of gcc (because the two compilers sometimes make slightly different
contraction choices). Fortunately, providing -ffp-contract=off to both
compilers reveals agreement. Older versions of gcc did not support
-ffp-contract, and if using such a version, you'll need to use the reference
outputs (which you should be able to do now).