Better errors for invalid --cc parameters.
When you specified something odd for --cc you'd get a really cryptic "Permission Denied" error when we go to check the compiler version. Do better checking on the --cc executable first, and error if it is not  1) executable and 2) a file.  Also, make the compile test function in the same way.