Adding C-based sin/cos functions to sphereflake
Standard libraries have slightly different results on different targets.
To make the output reproducible in any target we need some C-based
implementation that is simple enough to be the same on all archs.

We could create a common test-suite library for all those functions
(sin/cos, sqrt, rand, etc) and change all tests that were cooked to
accept bogus results to use them, but that's for another commit.