[Polly] Introduce a new fast LNT builder for Polly with isl+SCEV codegen
Running with 16 threads and just on iteration, we gain are fast reports
(< 10 min) on possible LNT failures. Reliable performance numbers are
not necessarily expected.

To make room for this builder we drop one of our CLooG code generation
builders, as the CLooG code generation is soon to be replaced anyhow.