Adapt daily report to handle runs missing on some days.
When a machine on a particular day doesn't produce a result, the daily
report page won't report regressions/improvements for the day after.
This patch fixes that, so that when a machine doesn't produce results
for a particular day, for what-ever reason, regressions and improvements
are still analyzed and reported.

Reasons for a machine not producing results every day could be the
machine having a hick-up for a day; or simply because that machine
is scheduled to not run a large benchmark every day, but e.g. only
every week.