[cmake] Add a very alpha driver for the CMake-based test-suite
This driver mimicks the usage of the 'nt' driver as much as possible, but only implements a subset of the parameters. The parameter list will surely grow as more people use this and discover their favourite option isn't implemented.

In its current state it can drive the test-suite for non-cross-builds, and collect compile and test times and submit them.

It cannot do test reruns yet. This is still very much a work in progress and reviews/feedback would be much appreciated.

Still missing:
  * support for manually setting ARCH.
  * support for cross-compilation.
  * support for using perf/timeit.sh.
  * support for using SMALL/LARGE datasets.
  * support for externals.
  * support for rununder.
  * ... and a ton more stuff that I can't think of now.

However, the line count is 21% of nt.py so far...