Update green dragon primary builds to CMake
This is the patch to update the stage 1 bot to use CMAKE instead of the
now deprecated auto tools it previously used. This patch addresses only
the stage 1 part of the build at the moment. The fix for stage 2 will
follow shortly.

This patch includes changes to build.py for the actual bot changes and
test_build.py to update the LIT tests to match. This should be a drop in
replacement for the current version with no changes required to the
Jenkins job.

Patch by Marc Schifer mschifer@apple.com